# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "User": {
        "properties": {
          "address": {
            "type": "string"
          },
          "emailOrPhone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "rate": {
            "type": "string"
          },
          "skills": {
            "type": "string"
          },
          "facebook": {
            "type": "string"
          },
          "linkedIn": {
            "type": "string"
          },
          "twitter": {
            "type": "string"
          },
          "instagram": {
            "type": "string"
          },
          "youtube": {
            "type": "string"
          },
          "telegram": {
            "type": "string"
          },
          "whatsapp": {
            "type": "string"
          },
          "tz": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "roles": {
            "items": {
              "enum": [
                "ROLE_USER"
              ],
              "type": "string"
            },
            "type": "array"
          },
          "id": {
            "minLength": 1,
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "updatedAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "emailOrPhone"
        ],
        "example": {
          "id": "c166ff1b-bc2c-4656-b363-d121075b8f6a"
        }
      },
...
      "UserSearchFilterDto": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "role": {
            "enum": [
              "ROLE_USER"
            ],
            "type": "string"
          }
        },
        "type": "object"
      },
      "UserSearchDto": {
        "properties": {
          "filter": {
            "$ref": "#/components/schemas/UserSearchFilterDto",
            "type": "object"
          },
          "page": {
            "minLength": 1,
            "type": "number"
          },
          "sort": {
            "type": "object"
          },
          "limit": {
            "type": "number"
          }
        },
        "type": "object",
        "required": [
          "filter",
          "page",
          "sort"
        ]
      },
...
      "User_search": {
        "type": "object",
        "description": "Subset of User serialized with class-transformer group \"search\".",
        "properties": {
          "id": {
            "minLength": 1,
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "updatedAt": {
            "oneOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "format": "date-time",
                "type": "string"
              }
            ]
          },
          "address": {
            "type": "string"
          },
          "emailOrPhone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "rate": {
            "type": "string"
          },
          "skills": {
            "type": "string"
          },
          "facebook": {
            "type": "string"
          },
          "linkedIn": {
            "type": "string"
          },
          "twitter": {
            "type": "string"
          },
          "instagram": {
            "type": "string"
          },
          "youtube": {
            "type": "string"
          },
          "telegram": {
            "type": "string"
          },
          "tz": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "roles": {
            "items": {
              "enum": [
                "ROLE_USER"
              ],
              "type": "string"
            },
            "type": "array"
          }
        }
      },
      "User_edit": {
        "type": "object",
        "description": "Subset of User serialized with class-transformer group \"edit\".",
        "properties": {
          "emailOrPhone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "company": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "rate": {
            "type": "string"
          },
          "skills": {
            "type": "string"
          },
          "facebook": {
            "type": "string"
          },
          "linkedIn": {
            "type": "string"
          },
          "twitter": {
            "type": "string"
          },
          "instagram": {
            "type": "string"
          },
          "youtube": {
            "type": "string"
          },
          "telegram": {
            "type": "string"
          },
          "tz": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "roles": {
            "items": {
              "enum": [
                "ROLE_USER"
              ],
              "type": "string"
            },
            "type": "array"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token (same value the API returns on login)"
      }
    }
  },
  "info": {
    "title": "API schema",
    "version": "v1"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/search": {
      "post": {
        "operationId": "UserController.search",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSearchDto",
                "type": "object",
                "required": [
                  "page",
                  "filter",
                  "sort"
                ],
                "properties": {
                  "page": {
                    "type": "integer",
                    "minimum": 0,
                    "description": "Zero-based page index"
                  },
                  "sort": {
                    "type": "object",
                    "additionalProperties": {
                      "type": "string"
                    },
                    "description": "Map of field name to sort direction (ASC or DESC)"
                  },
                  "limit": {
                    "type": "integer",
                    "description": "Optional page size override"
                  },
                  "filter": {
                    "type": "object",
                    "additionalProperties": true,
                    "description": "Endpoint-specific filters (see summary / examples)"
                  }
                }
              },
              "example": {
                "filter": {
                  "role": "ROLE_USER"
                },
                "sort": {
                  "createdAt": "ASC"
                },
                "page": 0
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "type": "array",
                        "items": {
                          "$ref": "#/components/schemas/User_search"
                        }
                      },
                      {
                        "type": "integer"
                      }
                    ]
                  }
                }
              }
            },
            "description": ""
          }
        },
        "summary": "Search users",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/user/{address}/address": {
      "get": {
        "operationId": "UserController.read",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User_search"
                }
              }
            },
            "description": ""
          },
          "404": {
            "description": "User does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "NotFoundError"
                    },
                    "message": {
                      "type": "string",
                      "example": "User does not exist"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Public profile by wallet address",
        "tags": [
          "User"
        ]
      }
    },
    "/api/user": {
      "put": {
        "operationId": "UserController.edit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User_edit"
              }
            }
          },
          "description": "",
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Update current user profile",
        "tags": [
          "User"
        ]
      }
    },
...
  }
}